requirements.txt
langchain
llama-index
chromadb
sentence-transformers
pypdf
streamlit
litellm
ragas
tiktoken
python-dotenv